+++ src/contexts/Socket/index.jsx
@@ -1,20 +1,8 @@
-/* eslint-disable no-unused-vars */
 import { createContext } from 'react';
-import SocketIOClient, { io } from 'socket.io-client';
+import { io } from 'socket.io-client';
 import { SERVICES_ENDPOINT } from 'constant';
 
-const socket = io(SERVICES_ENDPOINT);
-
-const SocketContext = createContext({
-  gameId: '',
-  hasGameStarted: false,
-  hasGameEnded: false,
-  isRoundActive: false,
-  isRoundCompleted: false,
-  setEndTurn: () => {},
-  shouldNotifyJoinGame: false,
-  notifyJoinGame: () => {},
-  socket,
-});
+export const socket = io(SERVICES_ENDPOINT);
+const SocketContext = createContext({ socket });
 
 export default SocketContext;
