+++ src/index.js
@@ -5,6 +5,7 @@ import './index.css';
 import { BrowserRouter } from 'react-router-dom';
 import ViewProvider from 'providers/View';
 import SocketProvider from 'providers/Socket';
+import RoundProvider from 'providers/Round';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
@@ -13,7 +14,9 @@ ReactDOM.render(
     <BrowserRouter>
       <ViewProvider>
         <SocketProvider>
-          <App />
+          <RoundProvider>
+            <App />
+          </RoundProvider>
         </SocketProvider>
       </ViewProvider>
     </BrowserRouter>
