+++ src/scenes/Leaderboard/index.jsx
@@ -1,5 +1,5 @@
 import { useContext, useEffect, useState } from 'react';
-import SocketContext from 'contexts/Socket';
+import RoundContext from 'contexts/Round';
 import InfoDialog from 'components/InfoDialog';
 import LeaderboardTable from 'scenes/Leaderboard/components/Table';
 import Content from 'content';
@@ -7,7 +7,7 @@ import useSWR from 'swr';
 import { fetcher } from 'services/utils';
 
 const LeaderboardPage = () => {
-  const { gameId } = useContext(SocketContext);
+  const { gameId } = useContext(RoundContext);
   const [leaderboardContent, setLeaderboardContent] = useState([]);
   const { data } = useSWR(`/api/v1/games/${gameId}/leaderBoard`, fetcher, {
     revalidateOnFocus: false,
